Modularize data types + logic
