+++ roles/test-utils/mining-device/src/lib/new_work_notifier.rs
@@ -1,0 +1,7 @@
+use async_channel::Sender;
+
+#[derive(Debug, Clone)]
+pub(crate) struct NewWorkNotifier {
+    pub(crate) should_send: bool,
+    pub(crate) sender: Sender<()>,
+}
