+++ roles/test-utils/mining-device/src/lib/next_share_outcome.rs
@@ -1,0 +1,10 @@
+pub(crate) enum NextShareOutcome {
+    ValidShare,
+    InvalidShare,
+}
+
+impl NextShareOutcome {
+    pub(crate) fn is_valid(&self) -> bool {
+        matches!(self, NextShareOutcome::ValidShare)
+    }
+}
