+++ roles/test-utils/mining-device/src/lib/setup_connection_handler.rs
@@ -1,0 +1,98 @@
+use super::{EitherFrame, StdFrame};
+use async_channel::{Receiver, Sender};
+use roles_logic_sv2::{
+    common_messages_sv2::{Protocol, SetupConnection, SetupConnectionSuccess},
+    handlers::common::ParseUpstreamCommonMessages,
+    parsers::MiningDeviceMessages,
+    routing_logic::{CommonRoutingLogic, NoRouting},
+    utils::Mutex,
+};
+use std::{convert::TryInto, net::SocketAddr, sync::Arc};
+use tracing::{error, info};
+
+pub(crate) struct SetupConnectionHandler {}
+
+impl SetupConnectionHandler {
+    pub fn new() -> Self {
+        SetupConnectionHandler {}
+    }
+    fn get_setup_connection_message(
+        address: SocketAddr,
+        device_id: Option<String>,
+    ) -> SetupConnection<'static> {
+        let endpoint_host = address.ip().to_string().into_bytes().try_into().unwrap();
+        let vendor = String::new().try_into().unwrap();
+        let hardware_version = String::new().try_into().unwrap();
+        let firmware = String::new().try_into().unwrap();
+        let device_id = device_id.unwrap_or_default();
+        info!(
+            "Creating SetupConnection message with device id: {:?}",
+            device_id
+        );
+        SetupConnection {
+            protocol: Protocol::MiningProtocol,
+            min_version: 2,
+            max_version: 2,
+            flags: 0b0000_0000_0000_0000_0000_0000_0000_0001,
+            endpoint_host,
+            endpoint_port: address.port(),
+            vendor,
+            hardware_version,
+            firmware,
+            device_id: device_id.try_into().unwrap(),
+        }
+    }
+    pub async fn setup(
+        self_: Arc<Mutex<Self>>,
+        receiver: &mut Receiver<EitherFrame>,
+        sender: &mut Sender<EitherFrame>,
+        device_id: Option<String>,
+        address: SocketAddr,
+    ) {
+        let setup_connection = Self::get_setup_connection_message(address, device_id);
+
+        let sv2_frame: StdFrame = MiningDeviceMessages::Common(setup_connection.into())
+            .try_into()
+            .unwrap();
+        let sv2_frame = sv2_frame.into();
+        sender.send(sv2_frame).await.unwrap();
+        info!("Setup connection sent to {}", address);
+
+        let mut incoming: StdFrame = receiver.recv().await.unwrap().try_into().unwrap();
+        let message_type = incoming.get_header().unwrap().msg_type();
+        let payload = incoming.payload();
+        ParseUpstreamCommonMessages::handle_message_common(
+            self_,
+            message_type,
+            payload,
+            CommonRoutingLogic::None,
+        )
+        .unwrap();
+    }
+}
+
+impl ParseUpstreamCommonMessages<NoRouting> for SetupConnectionHandler {
+    fn handle_setup_connection_success(
+        &mut self,
+        _: SetupConnectionSuccess,
+    ) -> Result<roles_logic_sv2::handlers::common::SendTo, roles_logic_sv2::errors::Error> {
+        use roles_logic_sv2::handlers::common::SendTo;
+        info!("Setup connection success");
+        Ok(SendTo::None(None))
+    }
+
+    fn handle_setup_connection_error(
+        &mut self,
+        _: roles_logic_sv2::common_messages_sv2::SetupConnectionError,
+    ) -> Result<roles_logic_sv2::handlers::common::SendTo, roles_logic_sv2::errors::Error> {
+        error!("Setup connection error");
+        todo!()
+    }
+
+    fn handle_channel_endpoint_changed(
+        &mut self,
+        _: roles_logic_sv2::common_messages_sv2::ChannelEndpointChanged,
+    ) -> Result<roles_logic_sv2::handlers::common::SendTo, roles_logic_sv2::errors::Error> {
+        todo!()
+    }
+}
